Rename cloud.ts -> httpClient.ts
