+++ app/src/lib/ai/service.ts
@@ -5,7 +5,7 @@ import { OpenAIModelName, type AIClient, AnthropicModelName } from '$lib/ai/type
 import { splitMessage } from '$lib/utils/commitMessage';
 import * as toasts from '$lib/utils/toasts';
 import OpenAI from 'openai';
-import type { CloudClient } from '$lib/backend/cloud';
+import type { CloudClient } from '$lib/backend/httpClient';
 import type { GitConfigService } from '$lib/backend/gitConfigService';
 import type { Hunk } from '$lib/vbranches/types';
