+++ app/src/routes/+layout.svelte
@@ -3,7 +3,7 @@
 
 	import { AIService } from '$lib/ai/service';
 	import { AuthService } from '$lib/backend/auth';
-	import { CloudClient, User } from '$lib/backend/cloud';
+	import { CloudClient, User } from '$lib/backend/httpClient';
 	import { GitConfigService } from '$lib/backend/gitConfigService';
 	import { ProjectService } from '$lib/backend/projects';
 	import { PromptService } from '$lib/backend/prompt';
